refactor object helper and database access, start to add caching where appropriate

+++ src/wbs/clients/apn/chat/broadcast/console/ChatBroadcastSendForm.java
@@ -22,7 +22,7 @@ class ChatBroadcastSendForm {
 	Gender gender;
 	Orient orient;
 
-	Integer categoryId;
+	Long categoryId;
 
 	Boolean hasPicture;
 	Boolean isAdult;
