+++ src/wbs/clients/apn/chat/report/console/ChatReportRevSharePart.java
@@ -391,7 +391,7 @@ class ChatReportRevSharePart
 				new ChatUserSearch ()
 
 			.chatId (
-				chat.getId ())
+				(long) chat.getId ())
 
 			.firstJoin (
 				TextualInterval.forInterval (
